Reserve sample app changes for a separate PR
